requirements/prod.txt
flask==1.1.1
flask-swagger-ui==3.20.9
gunicorn==20.0.4
python-json-logger==0.1.11
requests==2.22.0
PyYAML==5.2
